[Refactor] dataTime을 String으로 변환해주는 함수 DateFormatter의 타입 메소드로 리팩토링
